+++ lib/segment/src/vector_storage/quantized/quantized_chunked_mmap_storage/mod.rs
@@ -1,0 +1,5 @@
+mod read_only;
+mod read_write;
+
+pub use read_only::QuantizedChunkedStorageRead;
+pub use read_write::{QuantizedChunkedMmapStorage, QuantizedChunkedMmapStorageBuilder};
